Created status_test.go and TestComposeStatusString function
Modified composeStatusString to not take in a Status_Line object but instead numResources.
This makes testing this function easier, as we don't have to create a dummy list of resources. It only needed a Status_Line object to get the length of its resources field, anyway.

The new status_test.go file will hold the tests for status.go.
As of now, it only has one function TestComposeStatusString, which only has one test case.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>