Hid StatusLine fields behind methods
The values in the run method (e.g. bytesDownloaded, spinI, etc.) are now fields of the StatusLine struct.

lock.go no longer manipulates the StatusLine's internal fields. Instead, it calls NewStatusLine to get a new line object (and initialize and start the line) and calls increment when a resource finishes downloading.

The run method has been renamed to start and is now called inside NewStatusLine.

This encapsulation makes the StatusLine struct/class more modular. The NewStatusLine and increment methods act as an interface.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>