NewStatusLine now returns error, getResourceSizes & composeStatusLine now class methods, fixed crash & div-by-zero error
Fixed non-compiling code caused by the previous faulty merge.
Also fixed divide-by-zero runtime error when running TestDownload unit test. This was caused by totalBytes being less than the barLength (20 spaces). Now, when totalBytes is less than 20, the bar shortens (e.g. if totalBytes is 6, the bar will be 6 spaces wide).

getResourceSizes is now a class method. If fetching the resource sizes fails, it stores the http error in the new sizingErr field.
NewStatusLine calls getResourceSizes and then returns sizingErr (along with the pointer to the new line object). This potential error can then be handled by the caller -- as of now, if an error is returned, the Download function returns the error, terminating the program.

composeStatusLine is now also a class method. The sample test case in TestComposeStatusString has been removed.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>