Subscribed StatusLine to Download function's context
NewStatusLine now accepts a context parameter, which is stored in the new ctx field. In the StatusLine infinite for loop (in the start method), the blocking select statement has a new case that monitors the ctx (case <-st.ctx.Done(): return).

Now, the for loop will block until a value is inserted into indexCh (from completed resource or 50ms timer) OR the ctx is cancelled, which will end the loop and goroutine.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>