Download now continues even if sizing resources fails
If resource sizing fails, NewStatusLine will return the error for handling. Now, if an error is returned, Download does not call Start on the SL and sets the statusLine variable to nil. The download process will continue, but the created SL will not be updated or displayed.

If an SL's resource sizing fails, Start can still be called--the byte counter will be omitted and the bar will just show as [].

In both cases, an error message is printed to the terminal to tell the user why the SL is shortened or not being printed (in this case, we are not printing it).

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>