Implement gist
https://gist.github.com/rabadin/44a3d2a1f0c91e3f9e7c2b76440f3989

SL has a new field isRunning that is set to true in Start and set to false in the new Stop function. The SL loop calls Stop when the context is cancelled or when all resources are downloaded.

status_test.go has the new function TestSLWithCancelledContext.

The timeoutMs parameter has been removed from initResourceSizes and is a constant again.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>