Fix test fail
Adding a Sleep call in the Increment for loop in TestSLWith1000Resources seems to make the tests pass.

The last Assert has been changed to a Contains to exclude the variable second counter.
Each Sleep is now only 1 millisecond long instead of 10 -- this makes the test much faster while ensuring it still passes.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>