Fixed "elapsedd" bug
"elapsed" would eventually become "elapsedd". This occurred when the bytes-downloaded number eventually omitted decimals, shrinking the line width by one character. This would leave an extra 'd' at the end, as the current line is deleted by moving the cursor back and printing over it, as opposed to actually clearing it with "\033[2K", which causes flashing on Windows.

Fixed by adding a space at the end of every line printed.

Signed-off-by: Dylan Williams <mdylanwilliams576@gmail.com>